requirements.txt
PyMuPDF==1.23.24
pdfplumber==0.10.0
pyttsx3==2.90
gTTS==2.4.0
moviepy==1.0.3
streamlit==1.29.0
Pillow==10.1.0
requests==2.31.0
python-dotenv==1.0.0
argparse
pathlib
typing

# Enterprise AI/ML Platform Dependencies
pandas==2.1.4
numpy==1.24.3
scikit-learn==1.3.2
xgboost==2.0.3
lightgbm==4.1.0
prophet==1.1.5
plotly==5.17.0
altair==5.2.0
seaborn==0.13.0
matplotlib==3.8.2
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
pymongo==4.6.0
boto3==1.34.0
azure-storage-blob==12.19.0
google-cloud-storage==2.10.0
redis==5.0.1
joblib==1.3.2
dask==2023.12.1
streamlit-aggrid==0.3.4
streamlit-plotly-events==0.0.6
pyyaml==6.0.1
loguru==0.7.2
tenacity==8.2.3
pydantic==2.5.2
fastapi==0.104.1
uvicorn==0.24.0